ovn: Add ovn-nbctl.
ovn-nbctl is intended to be a utility for manually interacting with
the OVN northbound db.  A real consumer of OVN, such as OpenStack,
should be using ovsdb directly.  However, a utility like this is
useful during development and testing of both ovn itself as well as
the OpenStack Neutron plugin.  It could also be used when debugging
the state of an ovn deployment.

The commands related to logical switches and ports have been
implemented.  You can add, delete, list, and get/set external:ids.
On ports you can also get/set MAC addresses.

Signed-off-by: Russell Bryant <rbryant@redhat.com>
[blp@nicira.com adjusted C style and manpage fonts]
Signed-off-by: Ben Pfaff <blp@nicira.com>