openvswitch-fedora: remove requires to kernel
That RPM requires helped to work around a bug in
the kernel that was triggered by userspace package.

However, it doesn't force the kernel to be running,
so it's a weak solution.  Also, it makes difficult
to re-use the same spec on other RPM distros.

This patch comments out the requires.

Signed-off-by: Flavio Leitner <fbl@redhat.com>
Acked-by: Russell Bryant <rbryant@redhat.com>
Signed-off-by: Ben Pfaff <blp@nicira.com>